refactor(help-menu): expose interfaces in api to be used by backend
