+++ packages/api/src/help-menu.ts
@@ -1,0 +1,35 @@
+/**********************************************************************
+ * Copyright (C) 2025 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ***********************************************************************/
+
+export enum ActionKind {
+  LINK,
+  COMMAND,
+}
+
+export interface ItemAction {
+  kind: ActionKind;
+  parameter: string;
+}
+
+export interface ItemInfo {
+  title: string;
+  tooltip?: string;
+  icon: string;
+  enabled: boolean;
+  action?: ItemAction;
+}
